refactoring(service): token verification and error handling moved to axios interceptors
