--- client/src/shared/config/instance.ts
@@ -1,6 +1,0 @@
-import axios from "axios";
-import { BASE_URL } from "./url";
-
-export const $api = axios.create({
-  baseURL: BASE_URL,
-});
