+++ client/src/shared/config/saveTokens.ts
@@ -1,0 +1,6 @@
+import { AuthTokens } from "../../pages/Register/types";
+
+export const saveTokensLocalStoare = (data: AuthTokens): void => {
+  localStorage.setItem("accessToken", data.access.token);
+  localStorage.setItem("refreshToken", data.refresh.token);
+};
